Commit of not working crontab solution of docker.

+++ docker/toRoot/update-service.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 # Script to download deployment unit from a Maven artifact repository.
 
+APP=configservice.jar
+
 releaseRepo=http://mvnrepo.cantara.no/content/repositories/releases
 snapshotRepo=http://mvnrepo.cantara.no/content/repositories/snapshots
 groupId=no/cantara/jau
@@ -24,20 +26,39 @@ else #A specific Release version
    jarfile=$artifactId-$version.jar
 fi
 
-# Download artifact
-echo Downloading $url
-wget -O $jarfile -q -N $url
 
-# Create symlink or replace existing sym link
-if [ -h $artifactId.jar ]; then
-   unlink $artifactId.jar
+shaUrl=url.sha1
+shaFromWeb=$(wget $shaUrl -q -O -)
+if [ -f $APP ]; then
+  localSha=$(sha1sum $jarfile | awk '{print $1}')
+else
+  echo "No local app file found"
+  localSha=-1
 fi
-ln -s $jarfile $artifactId.jar
 
-# Delete old jar files
-jar=$artifactId*.jar
-nrOfJarFilesToDelete=`ls $jar -A1t | tail -n +6 | wc -l`
-if [[ $nrOfJarFilesToDelete > 0 ]]; then
-    echo Deleting $nrOfJarFilesToDelete old jar files. Keep the 4 newest + the symlink.
-    ls $jar -A1t | tail -n +6 | xargs rm
+if [ "$shaFromWeb" == "$localSha" ]; then
+  echo "Already got newest version. Not doing anything."
+else
+  echo "stopping running app, if it is running"
+  pkill -f '/usr/bin/java -Dlogback.configurationFile=./logback.xml -jar "$APP"&' # this doesn't work
+  echo Downloading $url
+  wget -O $jarfile -q -N $url
+
+  # Create symlink or replace existing sym link
+  if [ -h $artifactId.jar ]; then
+     unlink $artifactId.jar
+  fi
+  ln -s $jarfile $artifactId.jar
+
+  # Delete old jar files
+  jar=$artifactId*.jar
+  nrOfJarFilesToDelete=`ls $jar -A1t | tail -n +6 | wc -l`
+  if [[ $nrOfJarFilesToDelete > 0 ]]; then
+      echo Deleting $nrOfJarFilesToDelete old jar files. Keep the 4 newest + the symlink.
+      ls $jar -A1t | tail -n +6 | xargs rm
+  fi
+  
+  echo "Starting $APP"
+  /usr/bin/java -Dlogback.configurationFile=./logback.xml -jar "$APP"&
+
 fi
